refactor: move app resolving into its own helper
