+++ src/Plugin.php
@@ -5,13 +5,10 @@ use Illuminate\View\Engines\EngineResolver;
 use Illuminate\View\Engines\PhpEngine;
 use Illuminate\View\Factory;
 use Illuminate\View\FileViewFinder;
-use Orchestra\Testbench\Concerns\CreatesApplication;
 use Psalm\LaravelPlugin\ReturnTypeProvider\UrlReturnTypeProvider;
 use Psalm\Plugin\PluginEntryPointInterface;
 use Psalm\Plugin\RegistrationInterface;
 use SimpleXMLElement;
-use function file_exists;
-use function getcwd;
 use const DIRECTORY_SEPARATOR;
 use function unlink;
 use function dirname;
@@ -19,7 +16,6 @@ use function glob;
 
 class Plugin implements PluginEntryPointInterface
 {
-    use CreatesApplication;
 
     /** @var array<string> */
     public static $model_classes = [];
@@ -29,7 +25,7 @@ class Plugin implements PluginEntryPointInterface
      */
     public function __invoke(RegistrationInterface $registration, ?SimpleXMLElement $config = null)
     {
-        $app = $this->bootApp();
+        $app = ApplicationHelper::bootApp();
 
         $fake_filesystem = new FakeFilesystem();
 
@@ -195,78 +191,10 @@ class Plugin implements PluginEntryPointInterface
         return $factory;
     }
 
-    /**
-     * @param \Illuminate\Foundation\Application $app
-     */
-    protected function getEnvironmentSetUp($app): void
-    {
-        $app['config']->set('app.key', 'AckfSECXIvnK5r28GVIWUAxmbBSjTsmF');
-    }
-
     private function addOurStubs(RegistrationInterface $registration): void
     {
         foreach (glob(__DIR__ . '/Stubs/*.stubphp') as $stubFilePath) {
             $registration->addStubFile($stubFilePath);
         }
-    }
-
-    /**
-     * @return \Illuminate\Foundation\Application|\Laravel\Lumen\Application
-     */
-    private function bootApp()
-    {
-        if (file_exists($applicationPath = __DIR__.'/../../../../bootstrap/app.php')) { // Applications
-            $app = require $applicationPath;
-        } elseif (file_exists($applicationPath = getcwd().'/bootstrap/app.php')) { // Local Dev
-            $app = require $applicationPath;
-        } else { // Packages
-            $app = (new static)->createApplication();
-        }
-
-        if ($app instanceof \Illuminate\Contracts\Foundation\Application) {
-            $app->make(\Illuminate\Contracts\Console\Kernel::class)->bootstrap();
-        } elseif ($app instanceof \Laravel\Lumen\Application) {
-            $app->boot();
-        }
-
-        $app->register(\Barryvdh\LaravelIdeHelper\IdeHelperServiceProvider::class);
-
-        return $app;
-    }
-
-    /**
-     * Resolve application bootstrapper.
-     *
-     * @param  \Illuminate\Foundation\Application  $app
-     *
-     * @return void
-     */
-    protected function resolveApplicationBootstrappers($app)
-    {
-        // we want to keep the default psalm exception handler, otherwise the Laravel one will always return exit codes
-        // of 0
-        //$app->make('Illuminate\Foundation\Bootstrap\HandleExceptions')->bootstrap($app);
-        $app->make('Illuminate\Foundation\Bootstrap\RegisterFacades')->bootstrap($app);
-        $app->make('Illuminate\Foundation\Bootstrap\SetRequestForConsole')->bootstrap($app);
-        $app->make('Illuminate\Foundation\Bootstrap\RegisterProviders')->bootstrap($app);
-
-        $this->getEnvironmentSetUp($app);
-
-        $app->make('Illuminate\Foundation\Bootstrap\BootProviders')->bootstrap($app);
-
-        foreach ($this->getPackageBootstrappers($app) as $bootstrap) {
-            $app->make($bootstrap)->bootstrap($app);
-        }
-
-        $app->make('Illuminate\Contracts\Console\Kernel')->bootstrap();
-
-        $app['router']->getRoutes()->refreshNameLookups();
-
-        /**
-         * @psalm-suppress MissingClosureParamType
-         */
-        $app->resolving('url', static function ($url, $app) {
-            $app['router']->getRoutes()->refreshNameLookups();
-        });
     }
 }
